добавл решение по задачке с календарем и покрыл ее юнит тестами
